fix: remove play v1
