+++ src/components/base/PageHeader.vue
@@ -12,8 +12,8 @@
             <!-- nav -->
             <div class="basis-full grow">
                 <div class="flex flex-row pb-0 items-center justify-end gap-y-0 gap-x-7 my-0 pl-7 static w-auto top-6 duration-500 ease-in">
-                    <a class="font-medium sm:py-6 hover:underline decoration-2 cursor-pointer" @click="$router.push('/play')">play v1</a>
-                    <a class="font-medium sm:py-6 hover:underline decoration-2 cursor-pointer" @click="$router.push('/play2')">play v2</a>
+                    <a class="font-medium sm:py-6 hover:underline decoration-2 cursor-pointer" @click="$router.push('/play')">play</a>
+                    <!-- <a class="font-medium sm:py-6 hover:underline decoration-2 cursor-pointer" @click="$router.push('/play2')">play v2</a> -->
                     <!-- <a class="font-medium sm:py-6 hover:underline decoration-2 cursor-pointer" @click="openModal">more+</a> -->
                 </div>
             </div>
