+++ src/pages/PlayPage.vue
@@ -4,6 +4,8 @@ import numbersJson from "@/assets/data/numbers.json"
 import { storeToRefs } from 'pinia'
 import { useSound } from "@vueuse/sound"
 import numbersSfx from "@/assets/data/numbers.mp3"
+import correctSfx from "@/assets/audio/correct.mp3"
+import wrongSfx from "@/assets/audio/wrong.mp3"
 import audiowaveImg from "@/assets/img/audio_wave.png"
 import playImg from "@/assets/img/play.png"
 const currSprite = ref("")
@@ -22,7 +24,6 @@ const detectDeviceType = () => {
 }
 
 const spriteMap = numbersJson.reduce((obj, item) => {
-    console.log(item)
     obj[item.number] = item.sprite
     return obj
 }, {})
@@ -31,6 +32,9 @@ const { play } = useSound(numbersSfx, {
     sprite: spriteMap,
     interrupt: false
 })
+
+const { play: playCorrect } = useSound(correctSfx)
+const { play: playWrong } = useSound(wrongSfx)
 
 useHead({
   title: route.meta.title,
@@ -53,36 +57,39 @@ const getRandomSprite = () => {
 
 const checkGuess = (guessNumber: string) => {
     if (guessNumber == currSprite.value) {
-        console.log("yay")
+        playCorrect()
         incrementCorrect()
     } else {
-        console.log("nay")
+        playWrong()
         incrementMissed()
     }
+    showAnswer()
 }
 
 const makeGuess = (guessNumber: string) => {
     if (gameRunning.value) {
         checkGuess(guessNumber)
         currSprite.value = getRandomSprite()
-        play({id: currSprite.value})
+        setTimeout(() => {
+            play({id: currSprite.value})
+        }, 200)
         showHint.value = false
     }
 
 }
 
-// const toggleGameRun = debounce((event) => {
-//     if (event.code == "Space") {
-//         if (gameRunning.value) {
-//             // pause
-//             pause()
-//         } else {
-//             // start
-//             start()
-//         }
-//         showHint.value = false
-//     }
-// }, 100)
+const toggleGameRun = debounce((event) => {
+    if (event.code == "Space") {
+        if (gameRunning.value) {
+            // pause
+            pause()
+        } else {
+            // start
+            start()
+        }
+        showHint.value = false
+    }
+}, 100)
 
 const repeatAudio = () => {
     if (gameRunning.value) {
@@ -98,38 +105,27 @@ const handleShortcuts = debounce((event) => {
     }
 }, 100)
 
-onMounted(() => {
-    console.log("mount play")
-    if (detectDeviceType() == "Mobile") {
-        notCompatible.value = true
-    }
-    window.addEventListener("keypress", handleShortcuts)
-    // window.addEventListener("keydown", toggleGameRun)
+const gameRunning = ref(false)
+const start = () => {
+    if (gameRunning.value) return
+    currSprite.value = getRandomSprite()
+    gameRunning.value = true
     setTimeout(() => {
-        currSprite.value = getRandomSprite()
         play({id: currSprite.value})
-    }, 1000)
-})
+        focusInput()
+    }, 200)
+}
 
-onUnmounted(() => {
-    console.log("unmount play")
-    window.removeEventListener("keypress", handleShortcuts)
-    // window.removeEventListener("keydown", toggleGameRun)
-})
+const focusInput = () => {
+    if (guessInput.value) {
+        guessInput.value.focus()
+    }
+}
 
-const gameRunning = ref(true)
-// const start = () => {
-//     if (gameRunning.value) return
-//     console.log("start game")
-//     currSprite.value = getRandomSprite()
-    
-// }
-
-// const pause = () => {
-//     if (!gameRunning.value) return
-//     console.log("pause game")
-//     gameRunning.value = false
-// }
+const pause = () => {
+    if (!gameRunning.value) return
+    gameRunning.value = false
+}
 
 const hint = () => {
     const result = numbersJson.find(obj => obj.number == currSprite.value)
@@ -140,45 +136,108 @@ const hint = () => {
     }
 }
 
+const guessNumber = ref("")
+const guessInput = ref(null)
+
+const submit = () => {
+    makeGuess(guessNumber.value)
+    guessNumber.value = ""
+}
+
+const gameArea = ref(null)
+
+const answerVal = ref("answer")
+const answer = ref("?")
+const showAnswer = () => {
+    answer.value = currSprite.value
+    answerVal.value = guessNumber.value == currSprite.value ? "correct" : "missed"
+    setTimeout(() => {
+        answer.value = "?"
+        answerVal.value = "answer"
+    }, 500)
+}
+
+onMounted(() => {
+    if (detectDeviceType() == "Mobile") {
+        notCompatible.value = true
+    }
+    window.addEventListener("keypress", handleShortcuts)
+    window.addEventListener("keydown", toggleGameRun)
+    if (gameArea.value) {
+        gameArea.value.addEventListener("click", focusInput)
+    }
+})
+
+onUnmounted(() => {
+    window.removeEventListener("keypress", handleShortcuts)
+    window.removeEventListener("keydown", toggleGameRun)
+    if (gameArea.value) {
+        gameArea.value.removeEventListener("click", focusInput)
+    }
+})
+
+
 </script>
 
 <template>
     <div class="flex flex-col h-screen">
         <PageHeader />
-        <div v-if="notCompatible" class="max-w-[85rem] w-full mx-auto px-4 h-full flex flex-col py-4 sm:py-10">
-            <p class="text-xl text-center px-4 sm:px-2">Unfortunately, this web app cannot be used on a mobile browser. Please switch to a desktop browser to play.</p>
-        </div>
-        <div v-else class="max-w-[85rem] w-full mx-auto px-4 h-full flex flex-col py-4 sm:py-10">
-            <div class="p-2 sm:px-4">
-                <p class="font-light text-2xl text-center">do you really know numbers in cantonese?</p>
-                <p class="font-light text-xl text-center">type in what you hear</p>
-                <!-- <p class="text-center font-light">tap {{ gameRunning ? "pause" : "play" }} or press 'space' to {{ gameRunning ? "pause" : "start" }}</p> -->
-            </div>
-            <div v-if="gameRunning" class="flex-grow grid grid-cols-1 md:grid-cols-2 gap-6">
-                <div class="flex flex-col">
-                    <div class="flex-grow">
-                        <img class="w-2/3" :src="audiowaveImg" @click="repeatAudio"/>
+        <div ref="gameArea" class="max-w-[85rem] w-full mx-auto px-4 h-full flex flex-col py-4 sm:py-10">
+            <!-- instructions -->
+            <div v-if="!gameRunning">
+                <h1 class="text-5xl md:text-6xl py-4">Canto Numbers</h1>
+                <div class="grid md:grid-cols-2 gap-6">
+                    <!-- instructions -->
+                    <div class="text-xl md:text-2xl font-light">
+                        <p class="py-2"><span class="font-medium">About: </span>Learn to distinguish numbers in Cantonese by listening</p>
+                        <p class="py-2"><span class="font-medium">How to play: </span>Type in the number you hear, and submit it to see if the guess is correct or not</p>
+                        <p class="py-2"><span class="font-medium">Sounds: </span>Get aquainted to the correct and wrong sound effects</p>
+                        <div class="grid grid-cols-2 gap-6">
+                            <button class="my-4 px-8 py-2 bg-slate-100 hover:bg-slate-200 rounded-md" @click="playCorrect">correct</button>
+                            <button class="my-4 px-8 py-2 bg-slate-100 hover:bg-slate-200 rounded-md" @click="playWrong">wrong</button>
+                        </div>
                     </div>
-                    <p v-if="showHint" class="text-center text-xl font-medium py-2">{{ hint() }}</p>
-                    <p class="text-center font-light">tap wavelength or press 'r' to repeat</p>
-                </div>
-                <div class="flex flex-col">
-                    <div class="flex-grow">
-                        <NumberInput @submit="makeGuess"/>
+
+                    <!-- controls -->
+                    <div class="hidden md:block text-xl md:text-2xl font-light">
+                        <p class="py-2"><span class="font-medium">Controls: </span>Use the following keyboard controls for ease of use</p>
+                        <ul class="list-disc list-inside">
+                            <li>Space: Play / Pause Game</li>
+                            <li>Enter: Submit Guess</li>
+                            <li>R: Replay Audio</li>
+                            <li>H: Hint in Jyutping</li>
+                        </ul>
                     </div>
-                    <p class="text-center font-light">tap submit or press 'enter' to submit</p>
+                </div>
+                <div class="text-center">
+                    <button class="text-xl md:text-2xl font-light my-4 px-8 py-2 bg-slate-100 hover:bg-slate-200 rounded-md" @click="start">start <span class="hidden md:inline-block">(space)</span></button>
                 </div>
             </div>
-            <div v-else class="flex-grow mx-auto text-center">
-                <p class="">Paused</p>
-                <!-- <img :src="playImg" @click="start"/> -->
-            </div>
-            <div class="sm:pt-4">
-                <p class="text-center text-2xl font-medium my-4">Stats</p>
-                <div class="grid grid-cols-1 sm:grid-cols-3 text-center">
-                    <p><span class="font-medium">correct:</span> {{ correct }} / {{ total }}</p>
-                    <p><span class="font-medium">missed:</span> {{ missed }} / {{ total }}</p>
-                    <button @click="reset">reset</button>
+
+            <!-- game -->
+            <div v-else ref="gameArea" class="">
+                <div class="text-3xl font-medium py-4 grid grid-cols-1 md:grid-cols-2 gap-6 border-b-4 mb-8">
+                    <div class="grid grid-cols-2 gap-3 items-center">
+                        <p class="md:text-right">answer: </p>
+                        <p class="text-right md:text-left">{{ answer }}</p>
+                    </div>
+                    <div class="grid grid-cols-2 gap-3 items-center">
+                        <p class="md:text-right">hint:</p>
+                        <p class="text-right md:text-left">{{ showHint ? hint() : "" }}<button v-if="!showHint" class="bg-slate-100 hover:bg-slate-200 rounded-md px-2 font-light text-2xl" @click="showHint = true">show <span class="hidden md:inline-block">(h)</span></button></p>
+                    </div>
+                </div>
+                <div class="text-center">
+                    <input ref="guessInput" v-model.number="guessNumber" class="text-8xl text-center w-full [appearance:textfield] [&::-webkit-outer-spin-button]:appearance-none [&::-webkit-inner-spin-button]:appearance-none caret-transparent" type="number" placeholder="guess" autofocus @keyup.enter="submit"/>
+                </div>
+                <div class="grid grid-cols-3 gap-3 w-full py-4">
+                    <button class="text-2xl font-light text-center py-4 bg-slate-100 hover:bg-slate-200 rounded-md" @click="submit">submit <span class="hidden md:inline-block">(enter)</span></button>
+                    <button class="text-2xl font-light text-center py-4 bg-slate-100 hover:bg-slate-200 rounded-md" @click="repeatAudio">repeat <span class="hidden md:inline-block">(r)</span></button>
+                    <button class="text-2xl font-light text-center py-4 bg-slate-100 hover:bg-slate-200 rounded-md" @click="pause">pause <span class="hidden md:inline-block">(space)</span></button>
+                </div>
+                <div class="grid grid-cols-1 md:grid-cols-3 text-2xl font-light text-center items-center border-t-4 pt-8">
+                    <p class="py-1 md:py-0"><span class="font-medium">correct:</span> {{ correct }} / {{ total }}</p>
+                    <p class="py-1 md:py-0"><span class="font-medium">missed:</span> {{ missed }} / {{ total }}</p>
+                    <button class="bg-slate-100 hover:bg-slate-200 rounded-md py-4 mt-1 md:mt-0" @click="reset">reset stats</button>
                 </div>
             </div>
         </div>
